0-neural_style.py

+++ supervised_learning/0x0C-neural_style_transfer/0-neural_style.py
@@ -55,4 +55,4 @@ class NST:
         # print(image)
         image = tf.clip_by_value(image / 255, 0, 1)
         # print(image)
-        return image
+        return tf.cast(image, tf.float32)
